+++ data-migrations/2025-02-07-0310-add-column-to-questions.sql
@@ -1,0 +1,3 @@
+# This is a boolean to use the content from `sampleText` as the default answer
+ALTER TABLE `questions`
+ADD COLUMN `useSampleTextAsDefault` TINYINT(1) NOT NULL DEFAULT 0 AFTER `sampleText`;