+++ src/models/__tests__/Question.spec.ts
@@ -28,6 +28,7 @@ describe('Question', () => {
     requirementText: casual.sentences(3),
     guidanceText: casual.sentences(10),
     sampleText: casual.sentences(10),
+    useSampleTextAsDefault: true,
     displayOrder: casual.integer(1, 20),
   }
   beforeEach(() => {
@@ -40,6 +41,7 @@ describe('Question', () => {
     expect(question.requirementText).toEqual(questionData.requirementText);
     expect(question.guidanceText).toEqual(questionData.guidanceText);
     expect(question.sampleText).toEqual(questionData.sampleText);
+    expect(question.useSampleTextAsDefault).toEqual(questionData.useSampleTextAsDefault);
     expect(question.displayOrder).toEqual(questionData.displayOrder);
     expect(question.required).toEqual(false);
   });