+++ src/resolvers/question.ts
@@ -56,6 +56,7 @@ export const resolvers: Resolvers = {
       requirementText,
       guidanceText,
       sampleText,
+      useSampleTextAsDefault,
       required,
       questionOptions } }, context: MyContext): Promise<Question> => {
 
@@ -73,6 +74,7 @@ export const resolvers: Resolvers = {
             requirementText,
             guidanceText,
             sampleText,
+            useSampleTextAsDefault,
             required
           });
 
@@ -123,11 +125,13 @@ export const resolvers: Resolvers = {
     // update an existing question
     updateQuestion: async (_, { input: {
       questionId,
+      questionTypeId,
       displayOrder,
       questionText,
       requirementText,
       guidanceText,
       sampleText,
+      useSampleTextAsDefault,
       required,
       questionOptions } }, context: MyContext): Promise<Question> => {
 
@@ -154,6 +158,7 @@ export const resolvers: Resolvers = {
             requirementText: requirementText,
             guidanceText: guidanceText,
             sampleText: sampleText,
+            useSampleTextAsDefault: useSampleTextAsDefault,
             required: required,
             isDirty: questionData.isDirty
           });